chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/host-context-class-change-expected.txt

Class change affecting a node in a distributed node's subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(b, null).backgroundColor is transparent
FAIL getComputedStyle(b, null).backgroundColor should be rgb(0, 128, 0). Was rgba(0, 0, 0, 0).
PASS successfullyParsed is true

TEST COMPLETE
